fix: Windows build

+++ src/rust/src/libccxr_exports/time.rs
@@ -1,6 +1,5 @@
 use std::convert::TryInto;
-use std::ffi::CStr;
-use std::os::raw::{c_char, c_int, c_long};
+use std::ffi::{c_char, c_int, c_long, CStr};
 
 use crate::{
     bindings::*, cb_708, cb_field1, cb_field2, ccx_common_timing_settings as timing_settings,
@@ -303,7 +302,7 @@ unsafe fn write_back_from_timing_info() {
     total_frames_count = timing_info.total_frames_count.as_u64().try_into().unwrap();
     gop_time = write_gop_time_code(timing_info.gop_time);
     first_gop_time = write_gop_time_code(timing_info.first_gop_time);
-    fts_at_gop_start = timing_info.fts_at_gop_start.millis();
+    fts_at_gop_start = timing_info.fts_at_gop_start.millis() as c_long;
     gop_rollover = if timing_info.gop_rollover { 1 } else { 0 };
     timing_settings.disable_sync_check = if timing_info.timing_settings.disable_sync_check {
         1
